kinda works fully

+++ game/version1/asteroid_1.py
@@ -11,8 +11,18 @@ import player
 game_window = pyglet.window.Window(800, 800, caption = "Meteora v0.01")
 
 main_batch = pyglet.graphics.Batch()
+score = 0
+lives = 3
+score_label0 = pyglet.text.Label(text="Score: ", x=10, y=game_window.height-24, batch=main_batch)
+score_label = pyglet.text.Label(text=str(score), x=10, y=game_window.height-44, batch=main_batch)
 
-score_label = pyglet.text.Label(text="Score = 0", x=10, y=game_window.height-24, batch=main_batch)
+player_icons = load.player_lives(lives, batch=main_batch)
+
+def update_score(score, label): 
+   label.text = str(score)
+   label.draw()
+
+
 level_label = pyglet.text.Label(text="Meteor Game", x=game_window.width//2, y=game_window.height-24, anchor_x='center', batch=main_batch)
 
 player_ship = player.Player(x=400, y=400, batch=main_batch)
@@ -35,14 +45,11 @@ def on_draw():
 
 
 def update(dt):
-
+   
    to_add = []
-
-   for obj in game_objects:
-      obj.update(dt)
-      to_add.extend(obj.new_objects)
-      obj.new_objects = []
-
+   global score
+   global lives
+   global player_icons
 
    for i in range(len(game_objects)):
       for j in range(i+1, len(game_objects)):
@@ -53,11 +60,33 @@ def update(dt):
                obj_1.handle_collision_with(obj_2)
                obj_2.handle_collision_with(obj_1)
 
+   for obj in game_objects:
+      obj.update(dt)
+      to_add.extend(obj.new_objects)
+      obj.new_objects = []
+
+   game_objects.extend(to_add)
+
    for to_remove in [obj for obj in game_objects if obj.dead]:
+         
+      if to_remove.is_bullet == False:
+         score += 1
+         update_score(score, score_label)
+
+      if to_remove.is_player == True and lives > 0:
+         lives -= 1
+         to_remove.x=400
+         to_remove.y=400
+         to_remove.velocity_x = 0
+         to_remove.velocity_y = 0
+         to_remove.rotation = 0
+         to_remove.dead = False
+         player_icons = load.player_lives(lives, batch=main_batch)
+         return
       to_remove.delete()
       game_objects.remove(to_remove)
 
-   game_objects.extend(to_add)
+   
 
 if __name__ == '__main__':
    game_window.push_handlers(player_ship)
